Use a shared package for Tailwind configuration
